guile: Remove glibtoolize replacement for HEAD
This removes the replacement of libtoolize with glibtoolize when building guile --HEAD.
This originally was introduced to make autogen.sh use the libtool provided by brew (#34769).
Since then upstream itself added glibtoolize detection to autogen.sh, which makes the brew build fail, because it replaces glibtoolize with gglibtoolize which does not exist.

Closes Homebrew/homebrew#44404.

Signed-off-by: Dominyk Tiller <dominyktiller@gmail.com>